(PC-12493) api: delete unused webapp endpoints
