--- api/src/pcapi/infrastructure/repository/favorite/favorite_domain_converter.py
@@ -1,11 +1,0 @@
-from pcapi.core.users.models import Favorite
-from pcapi.domain.favorite.favorite import FavoriteDomain
-
-
-def to_domain(favorite_sql_entity: Favorite, booking: dict = None) -> FavoriteDomain:
-    return FavoriteDomain(
-        identifier=favorite_sql_entity.id,
-        offer=favorite_sql_entity.offer,
-        mediation=favorite_sql_entity.mediation,
-        booking=booking,
-    )
