+++ api/src/pcapi/routes/__init__.py
@@ -46,7 +46,6 @@ def install_routes(app: Flask) -> None:
     from . import pro
     from . import saml
     from . import shared
-    from . import webapp
 
     adage.install_routes(app)
     external.install_routes(app)
@@ -55,7 +54,6 @@ def install_routes(app: Flask) -> None:
     pro.install_routes(app)
     saml.install_routes(app)
     shared.install_routes(app)
-    webapp.install_routes(app)
     adage_iframe.install_routes(app)
 
     app.register_blueprint(private_api)
