+++ api/src/pcapi/routes/native/v1/serialization/account.py
@@ -6,6 +6,10 @@ from uuid import UUID
 
 from dateutil.relativedelta import relativedelta
 import flask
+from jwt import DecodeError
+from jwt import ExpiredSignatureError
+from jwt import InvalidSignatureError
+from jwt import InvalidTokenError
 import pydantic
 from pydantic.class_validators import root_validator
 from pydantic.class_validators import validator
@@ -31,6 +35,8 @@ from pcapi.core.users.models import User
 from pcapi.core.users.models import UserRole
 from pcapi.core.users.models import VOID_FIRST_NAME
 from pcapi.core.users.models import VOID_PUBLIC_NAME
+from pcapi.core.users.utils import decode_jwt_token
+from pcapi.core.users.utils import sanitize_email
 from pcapi.models.feature import FeatureToggle
 from pcapi.routes.native.utils import convert_to_cent
 from pcapi.serialization.utils import to_camel
@@ -119,6 +125,42 @@ class CallToActionMessage(BaseModel):
         alias_generator = to_camel
         allow_population_by_field_name = True
         use_enum_values = True
+
+
+class ChangeBeneficiaryEmailBody(BaseModel):
+    token: str
+
+
+class ChangeEmailTokenContent(BaseModel):
+    current_email: pydantic.EmailStr
+    new_email: pydantic.EmailStr
+
+    @classmethod
+    @validator("current_email,new_email", pre=True)
+    def validate_emails(cls, email: str) -> str:
+        try:
+            return sanitize_email(email)
+        except Exception as e:
+            raise ValueError(email) from e
+
+    @classmethod
+    def from_token(cls, token: str) -> "ChangeEmailTokenContent":
+        try:
+            jwt_payload = decode_jwt_token(token)
+        except (
+            ExpiredSignatureError,
+            InvalidSignatureError,
+            DecodeError,
+            InvalidTokenError,
+        ) as error:
+            raise InvalidTokenError() from error
+
+        if not {"new_email", "current_email"} <= set(jwt_payload):
+            raise InvalidTokenError()
+
+        current_email = jwt_payload["current_email"]
+        new_email = jwt_payload["new_email"]
+        return cls(current_email=current_email, new_email=new_email)
 
 
 class SubscriptionMessage(BaseModel):
