--- api/src/pcapi/routes/serialization/mailing_contacts_serialize.py
@@ -1,11 +1,0 @@
-from pcapi.routes.native.v1.serialization import BaseModel
-
-
-class MailingContactResponseModel(BaseModel):
-    pass
-
-
-class MailingContactBodyModel(BaseModel):
-    email: str
-    dateOfBirth: str
-    departmentCode: str
