--- api/src/pcapi/routes/webapp/music_types.py
@@ -1,11 +1,0 @@
-from flask import jsonify
-from flask_login import login_required
-
-from pcapi.domain.music_types import music_types
-from pcapi.routes.apis import private_api
-
-
-@private_api.route("/musicTypes", methods=["GET"])
-@login_required
-def list_music_types():
-    return jsonify(music_types), 200
