+++ api/src/pcapi/utils/includes.py
@@ -1,6 +1,3 @@
-import copy
-
-
 USER_INCLUDES = [
     "-culturalSurveyId",
     "-culturalSurveyFilledDate",
@@ -12,57 +9,5 @@ USER_INCLUDES = [
     "-validationToken",
     "hasPhysicalVenues",
 ]
-
-WEBAPP_GET_BOOKING_INCLUDES = [
-    "completedUrl",
-    "-displayAsEnded",
-    "isEventExpired",
-    "-educationalBookingId",
-    "-individualBookingId",
-    "-status",
-    {
-        "key": "stock",
-        "includes": [
-            "isBookable",
-            "isEventExpired",
-            "remainingQuantity",
-            "-dnBookedQuantity",
-            "-rawProviderQuantity",
-            "-educationalPriceDetail",
-            {
-                "key": "offer",
-                "includes": [
-                    "thumbUrl",
-                    "hasBookingLimitDatetimesPassed",
-                    "isBookable",
-                    "isDigital",
-                    "isEvent",
-                    "-idAtProvider",
-                    "-authorId",
-                    "-lastValidationDate",
-                    "-rankingWeight",
-                    "-dateUpdated",
-                    "-type",
-                    {
-                        "key": "stocks",
-                        "includes": [
-                            "isBookable",
-                            "isEventExpired",
-                            "remainingQuantity",
-                            "-dnBookedQuantity",
-                            "-rawProviderQuantity",
-                            "-educationalPriceDetail",
-                        ],
-                    },
-                    {"key": "venue", "includes": ["-validationToken", "-businessUnitId"]},
-                ],
-            },
-        ],
-    },
-    {"key": "mediation", "includes": ["thumbUrl"]},
-]
-
-WEBAPP_GET_BOOKING_WITH_QR_CODE_INCLUDES = copy.deepcopy(WEBAPP_GET_BOOKING_INCLUDES)
-WEBAPP_GET_BOOKING_WITH_QR_CODE_INCLUDES.append("qrCode")
 
 FEATURE_INCLUDES = ["nameKey"]
