+++ api/tests/core/users/test_repository.py
@@ -4,7 +4,6 @@ from datetime import datetime
 from freezegun import freeze_time
 import pytest
 
-from pcapi.core.bookings import factories as bookings_factories
 from pcapi.core.offerers import factories as offerers_factories
 from pcapi.core.offers import factories as offers_factories
 from pcapi.core.users import exceptions
@@ -12,7 +11,6 @@ from pcapi.core.users import factories as users_factories
 from pcapi.core.users import repository
 from pcapi.core.users.models import UserRole
 from pcapi.core.users.repository import get_users_with_validated_attachment_by_offerer
-from pcapi.domain.favorite.favorite import FavoriteDomain
 
 
 pytestmark = pytest.mark.usefixtures("db_session")
@@ -93,81 +91,6 @@ class GetNewlyEligibleUsersTest:
         assert set(users) == {user_just_18, user_just_18_ex_underage_beneficiary}
         users = repository.get_newly_eligible_age_18_users(since=date(2017, 12, 30))
         assert set(users) == {user_just_18, user_just_18_ex_underage_beneficiary, user_already_18}
-
-
-class FindByBeneficiaryTest:
-    def test_returns_a_list_of_beneficiary_favorites(self, app):
-        # given
-        beneficiary = users_factories.UserFactory()
-        venue = offers_factories.VenueFactory()
-        offer_1 = offers_factories.ThingOfferFactory(venue=venue)
-        mediation_1 = offers_factories.MediationFactory(offer=offer_1)
-        users_factories.FavoriteFactory(mediation=mediation_1, offer=offer_1, user=beneficiary)
-        offer_2 = offers_factories.ThingOfferFactory(venue=venue)
-        users_factories.FavoriteFactory(offer=offer_2, user=beneficiary)
-
-        # when
-        favorites = repository.find_favorites_domain_by_beneficiary(beneficiary.id)
-
-        # then
-        assert len(favorites) == 2
-        assert isinstance(favorites[0], FavoriteDomain)
-        assert isinstance(favorites[1], FavoriteDomain)
-
-    def test_should_not_return_favorites_of_other_beneficiary(self, app):
-        # given
-        beneficiary = users_factories.UserFactory()
-        other_beneficiary = users_factories.UserFactory()
-        offer = offers_factories.ThingOfferFactory()
-        mediation = offers_factories.MediationFactory(offer=offer)
-        users_factories.FavoriteFactory(mediation=mediation, offer=offer, user=other_beneficiary)
-
-        # when
-        favorites = repository.find_favorites_domain_by_beneficiary(beneficiary.id)
-
-        # then
-        assert len(favorites) == 0
-
-    def test_should_return_booking_when_favorite_offer_is_booked(self, app):
-        # given
-        beneficiary = users_factories.UserFactory()
-        venue = offers_factories.VenueFactory()
-        offer = offers_factories.ThingOfferFactory(venue=venue)
-        stock = offers_factories.StockFactory(offer=offer, price=0)
-        booking = bookings_factories.IndividualBookingFactory(stock=stock, individualBooking__user=beneficiary)
-        mediation = offers_factories.MediationFactory(offer=offer)
-        favorite = users_factories.FavoriteFactory(mediation=mediation, offer=offer, user=beneficiary)
-
-        # when
-        favorites = repository.find_favorites_domain_by_beneficiary(beneficiary.id)
-
-        # then
-        assert len(favorites) == 1
-        favorite = favorites[0]
-        assert favorite.is_booked is True
-        assert favorite.booking_identifier == booking.id
-        assert favorite.booked_stock_identifier == stock.id
-        assert favorite.booking_quantity == booking.quantity
-
-    def test_should_not_return_booking_when_favorite_offer_booking_is_cancelled(self, app):
-        # given
-        beneficiary = users_factories.UserFactory()
-        stock = offers_factories.StockFactory()
-        bookings_factories.CancelledBookingFactory(
-            stock=stock,
-            user=beneficiary,
-        )
-        offer = stock.offer
-        mediation = offers_factories.MediationFactory(offer=offer)
-        favorite = users_factories.FavoriteFactory(mediation=mediation, offer=offer, user=beneficiary)
-
-        # when
-        favorites = repository.find_favorites_domain_by_beneficiary(beneficiary.id)
-
-        # then
-        assert len(favorites) == 1
-        favorite = favorites[0]
-        assert favorite.is_booked is False
 
 
 class GetApplicantOfOffererUnderValidationTest:
